fix: perk group multipliers from traits not applying
This was a very difficult bug to track down. Apparently during hookExactClass mods_overrideField was creating a Deep Clone of the target perk tree table instead of adding it by reference. This resulted in the memory address of the table being different every time and therefore the weighted container contains() function was returning false to check for if the tree existed in the container. Converting this into a hook onto the create function keeps the memory address the same and everything works.