fix(msu): skills not being added when a skill with the same ID which is garbage exists in SkillsToAdd
This fixes the issue with characters with multiple skills that unequip and equip the weapon in their onAdded function causing the weapon's skills to be lost upon loading a saved game. This issue came up, for example, with the Old Swordmaster with the Juggernaut and Metzger perks being simultaneously present.

This commit is temporary and should be removed once MSU adds the fix directly.